+++ lib/src/subsystems/achievements/achievements.dart
@@ -1,0 +1,4 @@
+export 'achievement.dart';
+export 'achievement_events.dart';
+export 'achievement_manager.dart';
+export 'achievement_provider.dart';